Changed interface: node handles are now "long".
git-svn-id: svn://svn.code.sf.net/p/meddly/code/trunk@418 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/meddly_expert.h
@@ -2738,9 +2738,9 @@ class MEDDLY::binary_operation : public operation {
     // low-level front ends
 
     /// Low-level compute on nodes a and b, return result.
-    virtual int compute(int a, int b);
+    virtual long compute(long a, long b);
     /// Low-level compute at level k on nodes a and b, return result.
-    virtual int compute(int k, int a, int b);
+    virtual long compute(int k, long a, long b);
 
   protected:
     inline void operationCommutes() {
