+++ src/operations/difference.cc
@@ -45,7 +45,7 @@ class MEDDLY::diffr_mdd : public generic_binary_mdd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::diffr_mdd::diffr_mdd(const binary_opname* opcode, 
@@ -55,7 +55,7 @@ MEDDLY::diffr_mdd::diffr_mdd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::diffr_mdd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::diffr_mdd::checkTerminals(long a, long b, long& c)
 {
   if (a == 0 || b == -1) {
     c = 0;
@@ -97,7 +97,7 @@ class MEDDLY::diffr_mxd : public generic_binary_mxd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::diffr_mxd::diffr_mxd(const binary_opname* opcode, 
@@ -107,7 +107,7 @@ MEDDLY::diffr_mxd::diffr_mxd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::diffr_mxd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::diffr_mxd::checkTerminals(long a, long b, long& c)
 {
   if (a == 0) {
     c = 0;
