+++ src/operations/intersection.cc
@@ -45,7 +45,7 @@ class MEDDLY::inter_mdd : public generic_binary_mdd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::inter_mdd::inter_mdd(const binary_opname* opcode, 
@@ -55,7 +55,7 @@ MEDDLY::inter_mdd::inter_mdd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::inter_mdd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::inter_mdd::checkTerminals(long a, long b, long& c)
 {
   if (a == 0 || b == 0) {
     c = 0;
@@ -104,7 +104,7 @@ class MEDDLY::inter_mxd : public generic_binary_mxd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::inter_mxd::inter_mxd(const binary_opname* opcode, 
@@ -114,7 +114,7 @@ MEDDLY::inter_mxd::inter_mxd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::inter_mxd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::inter_mxd::checkTerminals(long a, long b, long& c)
 {
   if (a == 0 || b == 0) {
     c = 0;
